+++ plugin/brdrq/brdrq_module_importer.py
@@ -12,7 +12,7 @@ from PyQt5.QtWidgets import QMessageBox
 # https://github.com/qgis/QGIS/issues/45646
 
 
-brdr_version = "0.15.3"
+brdr_version = "0.15.4"
 
 
 def find_python():
@@ -28,30 +28,37 @@ def find_python():
 
     raise Exception("Python executable not found")
 
+def pipinstall_in_libs(python_exe,package):
+    plugin_dir = os.path.dirname(__file__)
+    target_dir = os.path.join(plugin_dir, "libs")
+
+    # Add target_dir to sys.path
+    if target_dir not in sys.path:
+        sys.path.append(target_dir)
+        subprocess.check_call(
+            [python_exe, "-m", "pip", "install", package,"--target",target_dir]
+        )
+
 def pipinstall_by_subprocess(python_exe,package):
     if sys.platform != "win32":
-        # user_profile_name = iface.userProfileManager().userProfile().name()
-        # target_dir = os.path.expanduser(
-        #     os.path.join(
-        #         "~/.local/share/QGIS/QGIS3/profiles", user_profile_name, "python"
-        #     )
-        # )
-        plugin_dir = os.path.dirname(__file__)
-        target_dir = os.path.join(plugin_dir, "libs")
-
-        # Add target_dir to sys.path
-        if target_dir not in sys.path:
-            sys.path.append(target_dir)
         try:
-            subprocess.check_call(
-                [python_exe, "-m", "pip", "install", package,"--target",target_dir]
-            )
+            pipinstall_in_libs(python_exe,package)
         except:
             QMessageBox.error(None, "Error", f"This plugin needs external dependency '{package}', automatically installed for Windows. For Linux/Mac, the correct version ({package}) has to be installed manually")
+
     else:
-        subprocess.check_call(
-            [python_exe, "-m", "pip", "install", package]
-        )
+        try:
+            subprocess.check_call(
+                [python_exe, "-m", "pip", "install", package]
+            )
+        except:
+            try:
+                subprocess.check_call(
+                    [python_exe, "-m", "pip", "install", "--user", package]
+                )
+            except:
+                pipinstall_in_libs(python_exe,package)
+
 
 def install_brdr(python_exe):
     if "brdr" in sys.modules: